+++ src/hiero_sdk_python/file/file_append_transaction.py
@@ -19,18 +19,12 @@ from typing import TYPE_CHECKING, Any
 from hiero_sdk_python.file.file_id import FileId
 from hiero_sdk_python.hapi.services import file_append_pb2
 from hiero_sdk_python.hapi.services.schedulable_transaction_body_pb2 import SchedulableTransactionBody
-from hiero_sdk_python.hapi.services import file_append_pb2
-from hiero_sdk_python.hapi.services.schedulable_transaction_body_pb2 import (
-    SchedulableTransactionBody,
-)
 from hiero_sdk_python.hbar import Hbar
 from hiero_sdk_python.transaction.chunked_transaction import ChunkedTransaction
-from hiero_sdk_python.transaction.transaction import Transaction
 
 
 if TYPE_CHECKING:
     from hiero_sdk_python.channels import _Channel
-    from hiero_sdk_python.client.client import Client
     from hiero_sdk_python.executable import _Method
 
 
@@ -68,9 +62,6 @@ class FileAppendTransaction(ChunkedTransaction):
             self.set_chunk_size(chunk_size)
 
         self._total_chunks = self._calculate_total_chunks()
-        # Internal tracking for chunking
-        self._current_chunk_index: int | None = None
-        self._total_chunks: int = self._calculate_total_chunks()
 
     def _encode_contents(self, contents: str | bytes | None) -> bytes | None:
         """
@@ -87,26 +78,6 @@ class FileAppendTransaction(ChunkedTransaction):
         if isinstance(contents, str):
             return contents.encode("utf-8")
         return contents
-
-    def _calculate_total_chunks(self) -> int:
-        """
-        Calculates the total number of chunks needed for the current contents.
-
-        Returns:
-            int: The total number of chunks needed.
-        """
-        if self.contents is None:
-            return 1
-        return math.ceil(len(self.contents) / self.chunk_size)
-
-    def get_required_chunks(self) -> int:
-        """
-        Gets the number of chunks required for the current contents.
-
-        Returns:
-            int: The number of chunks required.
-        """
-        return self._calculate_total_chunks()
 
     def set_file_id(self, file_id: FileId) -> FileAppendTransaction:
         """
@@ -138,31 +109,25 @@ class FileAppendTransaction(ChunkedTransaction):
         self._total_chunks = self._calculate_total_chunks()
         return self
 
-    def set_max_chunks(self, max_chunks: int) -> FileAppendTransaction:
+    def _calculate_total_chunks(self) -> int:
         """
-        Sets the maximum number of chunks allowed for this transaction.
-
-        Args:
-            max_chunks (int): The maximum number of chunks allowed.
+        Calculates the total number of chunks needed for the current contents.
 
         Returns:
-            FileAppendTransaction: This transaction instance.
+            int: The total number of chunks needed.
         """
-        super().set_max_chunks(max_chunks)
-        return self
+        if self.contents is None:
+            return 1
+        return math.ceil(len(self.contents) / self.chunk_size)
 
-    def set_chunk_size(self, chunk_size: int) -> FileAppendTransaction:
+    def get_required_chunks(self) -> int:
         """
-        Sets the chunk size for this transaction.
-
-        Args:
-            chunk_size (int): The size of each chunk in bytes.
+        Gets the number of chunks required for the current contents.
 
         Returns:
-            FileAppendTransaction: This transaction instance.
+            int: The number of chunks required.
         """
-        super().set_chunk_size(chunk_size)
-        return self
+        return self._calculate_total_chunks()
 
     def _build_proto_body(self) -> file_append_pb2.FileAppendTransactionBody:
         """
@@ -247,182 +212,6 @@ class FileAppendTransaction(ChunkedTransaction):
         self.contents = proto.contents
         self._total_chunks = self._calculate_total_chunks()
         return self
-
-    def _validate_chunking(self) -> None:
-        """
-        Validates that the transaction doesn't exceed the maximum number of chunks.
-
-        Raises:
-            ValueError: If the transaction exceeds the maximum number of chunks.
-        """
-        if self.max_chunks and self.get_required_chunks() > self.max_chunks:
-            raise ValueError(
-                f"Cannot execute FileAppendTransaction with more than {self.max_chunks} chunks. "
-                f"Required: {self.get_required_chunks()}"
-            )
-
-    def freeze_with(self, client: Client) -> FileAppendTransaction:
-        """
-        Freezes the transaction by building the transaction body and setting necessary IDs.
-
-        For multi-chunk transactions, this method generates multiple transaction IDs
-        with incremented timestamps based on the chunk interval.
-
-        Args:
-            client (Client): The client instance to use for setting defaults.
-
-        Returns:
-            FileAppendTransaction: The current transaction instance for method chaining.
-        """
-        if self._transaction_body_bytes:
-            return self
-
-        self._resolve_transaction_id(client)
-        self._resolve_node_ids(client)
-
-        required_chunks = self.get_required_chunks()
-        self._generate_transaction_ids(self._transaction_ids[0], required_chunks)
-
-        for chunk in range(required_chunks):
-            self._current_chunk_index = chunk
-            self._current_transaction_id_index = chunk
-
-            node_bytes = {}
-
-            for node_account_id in self.node_account_ids:
-                self._node_account_id = node_account_id
-
-                transaction_body = self.build_transaction_body()
-                transaction_body.transactionID.CopyFrom(self._transaction_ids[chunk]._to_proto())
-                transaction_body.nodeAccountID.CopyFrom(node_account_id._to_proto())
-
-                node_bytes[node_account_id] = transaction_body.SerializeToString()
-
-            self._transaction_body_bytes[self._transaction_ids[chunk]] = node_bytes
-
-        return super().freeze_with(client)
-
-    @overload
-    def execute(
-        self,
-        client: Client,
-        timeout: int | float | None = None,
-        wait_for_receipt: Literal[True] = True,
-        validate_status: bool = False,
-    ) -> TransactionReceipt: ...
-
-    @overload
-    def execute(
-        self,
-        client: Client,
-        timeout: int | float | None = None,
-        wait_for_receipt: Literal[False] = False,
-        validate_status: bool = False,
-    ) -> TransactionResponse: ...
-
-    def execute(
-        self,
-        client: Client,
-        timeout: int | float | None = None,
-        wait_for_receipt: bool = True,
-        validate_status: bool = False,
-    ) -> TransactionReceipt | TransactionResponse:
-        """
-        Executes the file append transaction.
-
-        For multi-chunk transactions, this method will execute all chunks sequentially and return first response.
-
-        Args:
-            client: The client to execute the transaction with.
-            timeout (int | float | None, optional): The total execution timeout (in seconds) for this execution.
-            wait_for_receipt (bool, optional): Whether to wait for consensus and return the receipt.
-                If False, the method returns a TransactionResponse immediately after submission.
-            validate_status: (bool):  Whether the query should automatically validate the transaction status (default False).
-
-        Returns:
-            TransactionReceipt: If wait_for_receipt is True (default)
-            TransactionResponse: If wait_for_receipt is False
-        """
-        # Return the first response (as per JavaScript implementation)
-        return self.execute_all(client, timeout, wait_for_receipt, validate_status)[0]
-
-    @overload
-    def execute_all(
-        self,
-        client: Client,
-        timeout: int | float | None = None,
-        wait_for_receipt: Literal[True] = True,
-        validate_status: bool = False,
-    ) -> list[TransactionReceipt]: ...
-
-    @overload
-    def execute_all(
-        self,
-        client: Client,
-        timeout: int | float | None = None,
-        wait_for_receipt: Literal[False] = False,
-        validate_status: bool = False,
-    ) -> list[TransactionResponse]: ...
-
-    def execute_all(
-        self,
-        client: Client,
-        timeout: int | float | None = None,
-        wait_for_receipt: bool = True,
-        validate_status: bool = False,
-    ) -> list[TransactionReceipt] | list[TransactionResponse]:
-        """
-        Executes the file append transaction.
-
-        This method will execute all chunks sequentially and return list of response for each chunked
-
-        Args:
-            client: The client to execute the transaction with.
-            timeout (int | float | None, optional): The total execution timeout (in seconds) for this execution.
-            wait_for_receipt (bool, optional): Whether to wait for consensus and return the receipt.
-                If False, the method returns a TransactionResponse immediately after submission.
-            validate_status: (bool):  Whether the query should automatically validate the transaction status (default False).
-
-        Returns:
-            List[TransactionReceipt]: If wait_for_receipt is True (default)
-            List[TransactionResponse]: If wait_for_receipt is False
-        """
-        self._validate_chunking()
-
-        if not self._transaction_body_bytes:
-            self.freeze_with(client)
-
-        # Single chunk transaction
-        if len(self._transaction_ids) == 1:
-            return [super().execute(client, timeout, wait_for_receipt, validate_status)]
-
-        # Multi-chunk transaction - execute all chunks
-        responses = []
-        self._current_transaction_id_index = 0
-        for index in range(len(self._transaction_ids)):
-            self._current_transaction_id_index = index
-            response = super().execute(client, timeout, wait_for_receipt, validate_status)
-            responses.append(response)
-
-        return responses
-
-    @property
-    def body_size_all_chunks(self) -> list[int]:
-        """Returns an array of body sizes for transactions with multiple chunks."""
-        self._require_frozen()
-        sizes = []
-
-        original_transaction_index = self._current_transaction_id_index
-
-        try:
-            for i, _ in enumerate(self._transaction_ids):
-                self._current_transaction_id_index = i
-                self._current_chunk_index = i
-                sizes.append(self.body_size)
-        finally:
-            self._current_transaction_id_index = original_transaction_index
-
-        return sizes
 
     @classmethod
     def _from_protobuf(cls, transaction_body):
